refactor(inference): separate macOS Metal + NEON out of the shared hot path
Relocate the per-matmul Metal offload (encoded row/rows, retained-block
hybrid transpose, FFN gate/up hybrid) and the inference-session lifecycle
calls into src/inference/metal_seam.rs, and the 14 Apple-Silicon NEON/dotprod
leaf kernels into src/inference/cpu_neon.rs. inference.rs's matmul hot path
now carries only cfg-gated dispatch seams, so surgical x86/CUDA work cannot
touch a macOS-only branch.

Pure source-structure change: zero binary/behaviour impact on any target —
the moved code was already cfg(target_os="macos")-excluded off macOS.
cpu_neon is gated on target_arch="aarch64" (not macOS) so it compiles on
aarch64-linux for CI verification; the dispatch sites stay macOS-gated so the
NEON is only *used* on macOS and behaviour is byte-identical.

Verified on Windows host: `cargo check` (x86/CUDA) and
`cargo check --target aarch64-unknown-linux-gnu` (compiles the NEON) both
clean; git diff shows no x86/scalar/CUDA logic changed; release-build CPU
decode byte-identical pre/post; EOL unchanged. The metal_seam::/cpu_neon::
dispatch call sites are macOS-cfg'd — confirm on a Mac / CI. inference.rs -543.
